+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,5 +1,6 @@
 package com.techcourse;
 
+import com.techcourse.servlet.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
